+++ packages/cli/src/index.ts
@@ -57,7 +57,7 @@ import earlyGetConfig from './util/get-config';
 import * as configFiles from './util/config/files';
 import getGlobalPathConfig from './util/config/global-path';
 import {
-  defaultAuthConfig,
+  getDefaultAuthConfig,
   defaultGlobalConfig,
 } from './util/config/get-default';
 import * as ERRORS from './util/errors-ts';
@@ -311,20 +311,10 @@ const main = async () => {
 
   let authConfig: AuthConfig;
   try {
-    authConfig = configFiles.readAuthConfigFile();
+    authConfig = configFiles.readAuthConfigFile(config);
   } catch (err: unknown) {
     if (isErrnoException(err) && err.code === 'ENOENT') {
-      authConfig = defaultAuthConfig;
-      try {
-        configFiles.writeToAuthConfigFile(authConfig);
-      } catch (err: unknown) {
-        output.error(
-          `An unexpected error occurred while trying to write the auth config to "${hp(
-            VERCEL_AUTH_CONFIG_PATH
-          )}" ${errorToString(err)}`
-        );
-        return 1;
-      }
+      authConfig = getDefaultAuthConfig();
     } else {
       output.error(
         `An unexpected error occurred while trying to read the auth config in "${hp(